requirements.txt
flask
transformers
torch
requests
python-dotenv
nltk
sentencepiece
